# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fraud Busters model-analysis and benchmark contract
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusExecutionMode = typeof HealthStatusExecutionMode[keyof typeof HealthStatusExecutionMode];


export const HealthStatusExecutionMode = {
  live: 'live',
  mock: 'mock',
  cached: 'cached',
  fallback: 'fallback',
  unavailable: 'unavailable',
} as const;

export interface HealthStatus {
  status: string;
  model_ready: boolean;
  execution_mode: HealthStatusExecutionMode;
  timestamp: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export interface FeatureOption {
  label: string;
  value: string;
}

export type FeatureDefinitionType = typeof FeatureDefinitionType[keyof typeof FeatureDefinitionType];


export const FeatureDefinitionType = {
  number: 'number',
  select: 'select',
  text: 'text',
} as const;

export interface FeatureDefinition {
  key: string;
  label: string;
  type: FeatureDefinitionType;
  /** @nullable */
  unit?: string | null;
  required: boolean;
  /** @nullable */
  min?: number | null;
  /** @nullable */
  max?: number | null;
  /** @nullable */
  placeholder?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  options?: FeatureOption[] | null;
}

/**
 * @nullable
 */
export type ModelConfigThresholdConfig = { [key: string]: unknown } | null;

export type ModelConfigMetadata = { [key: string]: unknown };

export interface ModelConfig {
  primary_model: string;
  supported_models: string[];
  feature_schema: FeatureDefinition[];
  /** @nullable */
  threshold_config?: ModelConfigThresholdConfig;
  metadata?: ModelConfigMetadata;
}

/**
 * Only the final model feature fields should be sent.
 */
export type AnalyzeInputTransaction = {[key: string]: string | number};

export interface AnalyzeInput {
  /** Only the final model feature fields should be sent. */
  transaction: AnalyzeInputTransaction;
}

export type BenchmarkRowStatus = typeof BenchmarkRowStatus[keyof typeof BenchmarkRowStatus];


export const BenchmarkRowStatus = {
  available: 'available',
  unavailable: 'unavailable',
  error: 'error',
} as const;

export interface BenchmarkRow {
  model: string;
  /** @nullable */
  roc_auc?: number | null;
  /** @nullable */
  precision?: number | null;
  /** @nullable */
  recall?: number | null;
  /** @nullable */
  f1?: number | null;
  status: BenchmarkRowStatus;
  /** @nullable */
  note?: string | null;
}

export type BenchmarkResponseMetadata = { [key: string]: unknown };

export type BenchmarkResponseExecutionMode = typeof BenchmarkResponseExecutionMode[keyof typeof BenchmarkResponseExecutionMode];


export const BenchmarkResponseExecutionMode = {
  live: 'live',
  mock: 'mock',
  cached: 'cached',
  fallback: 'fallback',
  unavailable: 'unavailable',
} as const;

export interface BenchmarkResponse {
  benchmark_table: BenchmarkRow[];
  metadata?: BenchmarkResponseMetadata;
  execution_mode: BenchmarkResponseExecutionMode;
  /** @nullable */
  timestamp?: string | null;
}

export type AnalyzeResponseExecutionMode = typeof AnalyzeResponseExecutionMode[keyof typeof AnalyzeResponseExecutionMode];


export const AnalyzeResponseExecutionMode = {
  live: 'live',
  mock: 'mock',
  cached: 'cached',
  fallback: 'fallback',
  unavailable: 'unavailable',
} as const;

/**
 * @nullable
 */
export type AnalyzeResponseTransaction = {[key: string]: string | number} | null;

/**
 * @nullable
 */
export type AnalyzeResponseMetadata = { [key: string]: unknown } | null;

export interface AnalyzeResponse {
  /**
     * @minimum 0
     * @maximum 1
     */
  fraud_score: number;
  model_used: string;
  execution_mode: AnalyzeResponseExecutionMode;
  /** @nullable */
  risk_band?: string | null;
  /** @nullable */
  transaction?: AnalyzeResponseTransaction;
  /** @nullable */
  benchmark_table?: BenchmarkRow[] | null;
  /** @nullable */
  metadata?: AnalyzeResponseMetadata;
  /** @nullable */
  signal_context?: string[] | null;
  /** @nullable */
  request_id?: string | null;
  /** @nullable */
  timestamp?: string | null;
}

export type OverviewResponseSystemStatus = typeof OverviewResponseSystemStatus[keyof typeof OverviewResponseSystemStatus];


export const OverviewResponseSystemStatus = {
  ready: 'ready',
  simulator: 'simulator',
  fallback: 'fallback',
  unavailable: 'unavailable',
} as const;

export type OverviewResponseExecutionMode = typeof OverviewResponseExecutionMode[keyof typeof OverviewResponseExecutionMode];


export const OverviewResponseExecutionMode = {
  live: 'live',
  mock: 'mock',
  cached: 'cached',
  fallback: 'fallback',
  unavailable: 'unavailable',
} as const;

export type OverviewResponseMetadata = { [key: string]: unknown };

export interface OverviewResponse {
  system_status: OverviewResponseSystemStatus;
  execution_mode: OverviewResponseExecutionMode;
  primary_model: string;
  /** @nullable */
  last_benchmark?: string | null;
  latest_analysis?: AnalyzeResponse;
  metadata?: OverviewResponseMetadata;
}