# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fraud Busters model-analysis and benchmark contract
 * OpenAPI spec version: 0.1.0
 */
import type { HealthStatusExecutionMode } from './healthStatusExecutionMode';

export interface HealthStatus {
  status: string;
  model_ready: boolean;
  execution_mode: HealthStatusExecutionMode;
  timestamp: string;
}